+++ doc/news.rst
@@ -6,9 +6,47 @@ Whats' new
 
 Changes, newest first :
 
-- 2018/09/19:
+- 2018/09/27:
+
+  - New operators:
+  
+    - ``ensemble_ts_plot`` (:doc:`scripts/ensemble_ts_plot`), with
+      python-user-friendly shortcut ``ts_plot``
+      ( :py:func:`~climaf.functions.ts_plot` ):
+      an alternative to ``curves`` (:doc:`scripts/curves`) for time
+      series, with more possibilities for customization
+    
+    - ``cLinearRegression`` (:doc:`scripts/cLinearRegression`):
+      computes linear regressions between two lon/lat/time datasets
+      (same size) or between a time series and a lon/lat/time
+
+    - ``curl_tau_atm`` (:doc:`scripts/curl_tau_atm`):to compute the
+      wind stress curl on regular grids (typically atmospheric grids)
+
+  - ``lonlatvert_interpolation``
+    (:py:func:`~climaf.functions.lonlatvert_interpolation`) replaces
+    ``lonlatpres_interpolation`` to make it more generic and suitable
+    for CliMAF
+
+  - Added :py:func:`~climaf.classes.crealms` to have a
+    mechanism of alias on the realms names (as for the frequencies) 
+
+  - when creating a portable atlas (directory containing the html index with the figures,
+    using the 'dirname' option of cell()), we now have an index file in the directory
+    listing all the figures of the atlas with their CRS. This allows any tool to parse
+    this index and filter the figures with keywords (in the CRS).
+
+  - Added :py:func:`~climaf.functions.cscalar` : this function
+    returns a float in python when applied on a CliMAF object that is
+    supposed to have one single value. It uses cMA to import a masked
+    array and returns only the float.
+    
   - Allow to choose log directory (for files climaf.log and last.out),
     using environment variable CLIMAF_LOG_DIR
+
+  - Bug fix on ds() for the access to daily datasets with the CMIP5 project
+
+======= end
 
 - 2017/05/02: